Funciona

+++ acciones.py
@@ -7,30 +7,17 @@ logger = Logger()
 
 def mover_todo(origen, destino,sobreescribir=False,dentro=False):
 
-    if comprobar_existe(destino) and not sobreescribir:
-        raise Exception(f"{destino} ya existe, -s para sobreescribir")
-
-    elif not comprobar_existe(origen):
-        raise Exception(f"{origen} no existe")
-
-    elif not dentro:
-
+    if dentro:
         for contenido in os.listdir(origen):
             if comprobar_existe(destino + contenido) and not sobreescribir:
                 raise Exception(f"{origen + contenido} ya existe, -s para sobreescribir")
-
             else:
-                if comprobar_existe(destino+contenido):
-                    os.remove(destino+contenido)
+                if comprobar_existe(destino + contenido):
+                    os.remove(destino + contenido)
                 shutil.move(origen + contenido, destino)
                 logger.log(f"Se ha movido {contenido} -> {destino}")
-
     else:
-        if comprobar_existe(destino):
-            os.remove(destino)
         shutil.move(origen,destino)
-        logger.log(f"Se ha movido {origen} -> {destino}")
-
 
 def mover(origen, destino,sobreescribir=False):
 
@@ -47,7 +34,8 @@ def mover(origen, destino,sobreescribir=False):
         raise Exception(f"{destino+ultimo} ya existe, -s para sobreescribir")
 
     else:
-        os.remove(destino+ultimo)
+        if comprobar_existe(destino + ultimo):
+            os.remove(destino+ultimo)
         shutil.move(origen,destino)
         logger.log(f"Se ha movido {origen} -> {destino}")
 
@@ -62,5 +50,3 @@ def renombrar(origen, destino,sobreescribir=False):
     else:
         os.rename(origen, destino)
         logger.log(f"{origen} -> {destino}")
-
-
